Remove `CyberArk` from verbiage to avoid clash
If I add this action to the GitHub Marketplace without changing the name, it will clash with the action under my username already present there.  I removed `CyberArk` from main verbiage so as to avoid this clashing and to allow us quick entry into the GitHub Marketplace.  At this time, I will mark my username's action as `DEPRECATED` and make the repository `ARCHIVED`.